moved TabIcon to Components folder
